made generic mouse wrapper for golog
